[IQDB] Switch to new implementation
