+++ .env.sample
@@ -32,7 +32,6 @@
 # export DANBOORU_APP_NAME=
 # export DANBOORU_VERSION=
 # export DANBOORU_HOSTNAME=
-# export DANBOORU_IQDBS_SERVER=
 
 #
 # Development Only
