+++ app/controllers/iqdb_queries_controller.rb
@@ -3,42 +3,6 @@ class IqdbQueriesController < ApplicationController
   before_action :detect_xhr, :throttle
 
   def show
-    if params[:v2].present?
-      new_version
-    else
-      old_version
-    end
-
-    respond_with(@matches) do |fmt|
-      fmt.html do |html|
-        html.xhr { render layout: false }
-      end
-
-      fmt.json do
-        render json: @matches, root: "posts"
-      end
-    end
-  rescue IqdbProxy::Error, IqdbProxyNew::Error => e
-    render_expected_error(500, e.message)
-  end
-
-  private
-
-  def old_version
-    if params[:file]
-      @matches = IqdbProxy.query_file(params[:file].tempfile)
-    elsif params[:url].present?
-      parsed_url = Addressable::URI.heuristic_parse(params[:url]) rescue nil
-      raise User::PrivilegeError, "Invalid URL" unless parsed_url
-      whitelist_result = UploadWhitelist.is_whitelisted?(parsed_url)
-      raise User::PrivilegeError, "Not allowed to request content from this URL" unless whitelist_result[0]
-      @matches = IqdbProxy.query(params[:url])
-    elsif params[:post_id]
-      @matches = IqdbProxy.query_path(Post.find(params[:post_id]).preview_file_path)
-    end
-  end
-
-  def new_version
     if params[:file]
       @matches = iqdb_proxy(:query_file, params[:file].tempfile)
     elsif params[:url].present?
@@ -52,10 +16,24 @@ class IqdbQueriesController < ApplicationController
     elsif params[:hash]
       @matches = iqdb_proxy(:query_hash, params[:hash])
     end
+
+    respond_with(@matches) do |fmt|
+      fmt.html do |html|
+        html.xhr { render layout: false }
+      end
+
+      fmt.json do
+        render json: @matches, root: "posts"
+      end
+    end
+  rescue IqdbProxy::Error => e
+    render_expected_error(500, e.message)
   end
 
+  private
+
   def iqdb_proxy(method, value)
-    IqdbProxyNew.send(method, value, params[:score_cutoff])
+    IqdbProxy.send(method, value, params[:score_cutoff])
   end
 
   def throttle
