+++ app/jobs/iqdb_remove_job_new.rb
@@ -2,6 +2,6 @@ class IqdbRemoveJobNew < ApplicationJob
   queue_as :iqdb_new
 
   def perform(post_id)
-    IqdbProxyNew.remove_post(post_id)
+    IqdbProxy.remove_post(post_id)
   end
 end
