+++ app/models/post.rb
@@ -1545,7 +1545,7 @@ class Post < ApplicationRecord
 
     module ClassMethods
       def iqdb_enabled?
-        Danbooru.config.iqdbs_server.present?
+        Danbooru.config.iqdb_server.present?
       end
 
       def remove_iqdb(post_id)
