+++ test/functional/iqdb_queries_controller_test.rb
@@ -3,7 +3,7 @@ require 'test_helper'
 class IqdbQueriesControllerTest < ActionDispatch::IntegrationTest
   context "The iqdb controller" do
     setup do
-      Danbooru.config.stubs(:iqdbs_server).returns("https://karasuma.donmai.us")
+      Danbooru.config.stubs(:iqdb_server).returns("https://karasuma.donmai.us")
       @user = create(:user)
       as(@user) do
         @posts = create_list(:post, 2)
@@ -24,7 +24,7 @@ class IqdbQueriesControllerTest < ActionDispatch::IntegrationTest
         end
 
         should "render a response" do
-          IqdbProxy.expects(:query).with(@url).returns(@mocked_response)
+          IqdbProxy.expects(:query_url).with(@url, nil).returns(@mocked_response)
           get_auth iqdb_queries_path(variant: "xhr"), @user, params: @params
           assert_select("#post_#{@posts[0].id}")
         end
@@ -41,8 +41,8 @@ class IqdbQueriesControllerTest < ActionDispatch::IntegrationTest
           }]
         end
 
-        should "redirect to iqdbs" do
-          IqdbProxy.expects(:query_path).with(@posts[0].preview_file_path).returns(@mocked_response)
+        should "redirect to iqdb" do
+          IqdbProxy.expects(:query_post).with(@posts[0], nil).returns(@mocked_response)
           get_auth iqdb_queries_path, @user, params: @params
           assert_select("#post_#{@posts[0].id}")
         end
